# [chore] test prod terser config