as addition, we could remove `unwrap` and `expect` functions for prod.

https://astexplorer.net/#/gist/7d94e28a646e052ff8c6d1df72cc69d1/adcac2a89c58847b7ddd302154e41c224aa7f27f

```
     Benchmark Results Summary    

duration phase no difference [-94ms to 109ms]
renderEnd phase estimated improvement -2ms [-4ms to 0ms] OR -2.49% [-5.54% to -0.3%]
render1000Items1End phase estimated improvement -15ms [-24ms to -1ms] OR -1.18% [-1.89% to -0.05%]
clearItems1End phase no difference [-1ms to 3ms]
render1000Items2End phase no difference [-8ms to 10ms]
clearItems2End phase no difference [-1ms to 0ms]
render5000Items1End phase no difference [-9ms to 53ms]
clearManyItems1End phase estimated improvement -3ms [-6ms to 0ms] OR -1.22% [-2.45% to -0.21%]
render5000Items2End phase no difference [-36ms to 2ms]
clearManyItems2End phase no difference [-3ms to 1ms]
render1000Items3End phase no difference [-1ms to 15ms]
append1000Items1End phase no difference [-16ms to 0ms]
append1000Items2End phase no difference [-7ms to 3ms]
updateEvery10thItem1End phase no difference [0ms to 4ms]
updateEvery10thItem2End phase no difference [0ms to 1ms]
selectFirstRow1End phase no difference [0ms to 0ms]
selectSecondRow1End phase no difference [0ms to 1ms]
removeFirstRow1End phase no difference [-2ms to 1ms]
removeSecondRow1End phase no difference [-4ms to 0ms]
swapRows1End phase no difference [-3ms to 2ms]
swapRows2End phase no difference [0ms to 7ms]
clearItems4End phase no difference [-4ms to 3ms]
paint phase no difference [0ms to 0ms]
```

<img width="770" alt="image" src="https://github.com/glimmerjs/glimmer-vm/assets/1360552/b8194c11-8c06-4c31-a56d-83e760ea628e">

[artifact-1.pdf](https://github.com/glimmerjs/glimmer-vm/files/13762854/artifact-1.pdf)